+++ src/main/factories/controllers/signup/signup-controller-factory.ts
@@ -1,0 +1,11 @@
+import { SignUpController } from '../../../../presentation/controllers/signup/signup-controller'
+import { Controller } from '../../../../presentation/protocols'
+import { makeSignUpValidation } from './signup-validation-factory'
+import { makeDbAuthentication } from '../../usecases/authentication/db-authentication-factory'
+import { makeDbAddAccount } from '../../usecases/AddAccount/db-add-account-factory'
+import { makeLogControllerDecorator } from '../../decorators/log-controller-decorator-factory'
+
+export const makeSignUpController = (): Controller => {
+  const controller = new SignUpController(makeDbAddAccount(), makeSignUpValidation(), makeDbAuthentication())
+  return makeLogControllerDecorator(controller)
+}
