--- src/main/factories/login/login-factory.ts
@@ -1,21 +1,0 @@
-import env from '../../config/env'
-import { DbAuthentication } from '../../../data/usecases/authentication/db-authentication'
-import { BcryptAdapter } from '../../../infra/criptography/bcrypt-adapter/bcrypt-adapter'
-import { JwtAdapter } from '../../../infra/criptography/jwt-adapter/jwt-adapter'
-import { AccountMongoRepository } from '../../../infra/db/mongodb/account/account-mongo-repository'
-import { LogMongoRepository } from '../../../infra/db/mongodb/log/log-ongo-repository'
-import { LoginController } from '../../../presentation/controllers/login/login-controller'
-import { Controller } from '../../../presentation/protocols'
-import { LogControllerDecorator } from '../../decorators/log-controller-decorator'
-import { makeLoginValidation } from './login-validation-factory'
-
-export const makeLoginController = (): Controller => {
-  const salt = 12
-  const accountMongoRepository = new AccountMongoRepository()
-  const bcryptAdapter = new BcryptAdapter(salt)
-  const jwtAdapter = new JwtAdapter(env.jwtSecret)
-  const dbAuthentication = new DbAuthentication(accountMongoRepository, bcryptAdapter, jwtAdapter, accountMongoRepository)
-  const logMongoRepository = new LogMongoRepository()
-  const loginController = new LoginController(dbAuthentication, makeLoginValidation())
-  return new LogControllerDecorator(loginController, logMongoRepository)
-}
