+++ src/main/factories/usecases/authentication/db-authentication-factory.ts
@@ -1,0 +1,14 @@
+import env from '../../../config/env'
+import { DbAuthentication } from '../../../../data/usecases/authentication/db-authentication'
+import { BcryptAdapter } from '../../../../infra/criptography/bcrypt-adapter/bcrypt-adapter'
+import { JwtAdapter } from '../../../../infra/criptography/jwt-adapter/jwt-adapter'
+import { AccountMongoRepository } from '../../../../infra/db/mongodb/account/account-mongo-repository'
+import { Authentication } from '../../../../domain/usecases/authentication'
+
+export const makeDbAuthentication = (): Authentication => {
+  const salt = 12
+  const accountMongoRepository = new AccountMongoRepository()
+  const bcryptAdapter = new BcryptAdapter(salt)
+  const jwtAdapter = new JwtAdapter(env.jwtSecret)
+  return new DbAuthentication(accountMongoRepository, bcryptAdapter, jwtAdapter, accountMongoRepository)
+}
